requirements.txt
typer
textual
textual-dev
pycryptodome